requirements.txt
google-generativeai==0.8.5
langgraph==0.5.4
chainlit==2.6.2